Update step_1_setting_up_env_and_running_benchmark_from_samples.rst
The documentation makes some assumptions about the environment that Rally is running on top of that are not necessarily correct. Differences in the environment from those assumptions will lead to immediate failure of the first simple test that might discourage potential users of the tool. Further, Rally documentation does not currently appear to describe its internal operations down to the level of the temporary interactions with OpenStack, including creation of temporary tenants.